Added "StaffController" to which i moved some routes.
Also some authentication bug fixes
