+++ routes/web.php
@@ -1,10 +1,11 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\PagesController;
+use App\Http\Controllers\Auth\LoginController;
+use App\Http\Controllers\Auth\RegisterController;
 use App\Http\Controllers\FormController;
-use App\Http\Controllers\RegisterController;
-use App\Http\Controllers\LoginController;
+use App\Http\Controllers\PagesController;
+use App\Http\Controllers\StaffController;
+use Illuminate\Support\Facades\Route;
 
 /*
 |--------------------------------------------------------------------------
@@ -27,9 +28,6 @@ Route::controller(PagesController::class)->group(function (){
     Route::get('/aboutus', 'showAboutUs')->name('aboutUs');
     Route::get('/rent', 'showRent')->name('rent');
     Route::get('/termsofuse', 'showTermsofUse')->name('termsofUse');
-    Route::get('/crm', 'showCrm')->name('crm');
-    Route::get('/crm/{clientId}', 'showClientProfile')->name('clientProfile');
-    Route::get('/crm/{clientId}/edit', 'editClientProfile')->name('editClient');
     Route::get('/forms/iconsultform', 'showIconsultForm')->name('iconsultForm');
     Route::get('/messagepage', 'showMessage')->name('messagePage');
 });
@@ -40,13 +38,28 @@ Route::controller(FormController::class)->group(function (){
     Route::post('/forms/{clientId}/deleteClient', 'deleteClient')->name('deleteClient');
 });
 
-Route::controller(RegisterController::class)->group(function (){
+// Эту группу отличает использование посредника 'auth' Он позволяет открывать эти страницы только авторизованным пользователям.
+// в случае, если авторизации нет, то посредник сделает редирект на маршрут loginForm.
+Route::controller(StaffController::class)->middleware('auth')->group(function (){
+    Route::get('/crm', 'showCrm')->name('crm');
+    Route::get('/crm/{clientId}', 'showClientProfile')->name('clientProfile');
+    Route::get('/crm/{clientId}/edit', 'editClientProfile')->name('editClient');
+});
+
+// Посредники можно применять к группам маршрутов. Их действие будет распостраняться на все маршруты в группе.
+// Если в аргумент к middleware() объявить массив, то можно будет использовать сразу несколько посредников.
+Route::controller(RegisterController::class)->middleware('guest')->group(function (){
     Route::get('/register', 'showIndex')->name('registerForm');
     Route::post('/register', 'register')->name('registerAction');
 });
 
-Route::controller(LoginController::class)->group(function (){
+// Функция middleware('имя посредника') вызывает посредник из app/Http/Kernel.php. 'guest'- это алиас, который вызывает экземпляр класса
+// \App\Http\Middleware\RedirectIfAuthenticated. Его внутренний метод handle() проверяет, залогинен ли пользователь. Если нет, то он дает
+// маршруту сработать. Если да, то делает редирект константой HOME. Она прописывается в Providers/RouteServiceProvider. У меня назначен адрес '/'.
+Route::controller(LoginController::class)->middleware('guest')->group(function (){
     Route::get('/login', 'showIndex')->name('loginForm');
     Route::post('/login', 'login')->name('loginAction');
-    Route::post('/logout', 'logout')->name('logoutAction');
+// Метод withoutMiddleware('имя посредника') позволяет проигнорировать выбранный посредник, когда его действие распостраняется
+// на всю группу.
+    Route::post('/logout', 'logout')->name('logoutAction')->withoutMiddleware('guest');
 });
